Move WPMediaUtils.java to the main utils folder
