+++ WordPress/src/main/java/org/wordpress/android/ui/photopicker/PhotoPickerActivity.java
@@ -17,7 +17,7 @@ import android.view.MenuItem;
 import org.wordpress.android.BuildConfig;
 import org.wordpress.android.R;
 import org.wordpress.android.ui.RequestCodes;
-import org.wordpress.android.ui.media.WordPressMediaUtils;
+import org.wordpress.android.util.WordPressMediaUtils;
 import org.wordpress.android.ui.photopicker.PhotoPickerFragment.PhotoPickerOption;
 import org.wordpress.android.util.AppLog;
 import org.wordpress.passcodelock.AppLockManager;
